requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.4.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==2.3.6
